data/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anomalib==2.2.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
asttokens==3.0.1
attrs==25.4.0
bitsandbytes==0.49.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.19
decorator==5.2.1
dnspython==2.8.0
docstring-parser==0.17.0
einops==0.8.1
email-validator==2.3.0
executing==2.2.1
fastapi==0.135.2
fastapi-cli==0.0.24
fastapi-cloud-cli==0.15.1
fastar==0.9.0
filelock==3.20.3
fonttools==4.61.1
freia==0.2
frozenlist==1.8.0
fsspec==2026.1.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==1.3.2
hydra-core==1.3.2
idna==3.11
imagecodecs==2026.1.14
imageio==2.37.2
importlib-resources==6.5.2
ipykernel==7.1.0
ipython==9.9.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jinja2==3.1.6
joblib==1.5.3
jsonargparse==4.45.0
jsonnet==0.21.0
jupyter-client==8.8.0
jupyter-core==5.9.1
kiwisolver==1.4.9
kornia==0.8.2
kornia-rs==0.1.10
lazy-loader==0.4
lightning==2.6.0
lightning-utilities==0.15.2
markdown-it-py==4.0.0
markupsafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==12.1.0
platformdirs==4.5.1
prompt-toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.4
psutil==7.2.1
ptyprocess==0.7.0
pure-eval==0.2.3
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-extra-types==2.11.1
pydantic-settings==2.13.1
pygments==2.19.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
pytorch-lightning==2.6.0
pytz==2025.2
pyyaml==6.0.3
pyzmq==27.1.0
requests==2.32.5
rich==14.2.0
rich-argparse==1.7.2
rich-toolkit==0.19.7
rignore==0.7.6
safetensors==0.7.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.0
sentry-sdk==2.56.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
stack-data==0.6.3
starlette==1.0.0
sympy==1.14.0
tensorboardx==2.6.4
threadpoolctl==3.6.0
tifffile==2026.1.14
timm==1.0.24
torch==2.9.1
torchmetrics==1.8.2
torchvision==0.24.1
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
triton==3.5.1
typer==0.24.1
typer-slim==0.21.1
typeshed-client==2.8.2
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.42.0
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.2.14
websockets==16.0
yarl==1.22.0
